refactor: split cmap.go into font.go (Font public API) and cmap.go (CMap internals)
font.go: Font struct + 7 methods (BaseFont, FirstChar, LastChar, Widths,
Width, Encoder, getEncoder). No behaviour change; pure move.

cmap.go: retains CMap parsing internals — dictEncoder, byteRange, bfchar,
bfrange, cmap, cmapInterp, readCmap and their methods.

Separation makes the two responsibilities explicit: Font is a public value
type whose methods read a PDF font dictionary; CMap parsing is an internal
detail of how Encoder() resolves a ToUnicode stream into a TextEncoding.
